Linting - Addition of custom Types for props
Exceptions were added at the web3.ts file for the functions that are not in use.

The `DepositCard` processing logic of the balance expected a number and the component instantiations at the `DepositList` were assigned a string value. The string values were parsed to floats to match the types on `DepositList.tsx`