QA Linting - Removal of unused variables, non-null assertions and React-Hook warnings
This commit resolves all the eslint warnings and errors left.

Multiple variables, modules, functions and props were removed as they were not in use.

Warnings for non-null assertions were disabled in few locations with their respective reasons described in the comments. In other instances, where the props 'non-nulled' were found to always have a value, the non-null assertion was removed.

Warnings for the 'react-hooks/exhaustive-deps' eslint rule were disabled. This warnings were generated by the use of useEffect as a trigger of function upon a variable change or on mount.

Custom props and return types were created for the reduceGraphResult() method at contractReducers.ts.